Extract even more `_blocking` function flavors to a dedicated file
