+++ hermes/bin/src/runtime_extensions/hermes/ipfs/host.rs
@@ -4,11 +4,10 @@ use hermes_ipfs::Cid;
 
 use crate::{
     ipfs::{
-        self, hermes_ipfs_add_file, hermes_ipfs_content_validate, hermes_ipfs_dht_get_providers,
-        hermes_ipfs_dht_provide, hermes_ipfs_evict_peer, hermes_ipfs_get_dht_value,
-        hermes_ipfs_get_file, hermes_ipfs_get_peer_identity_blocking, hermes_ipfs_pin_file,
-        hermes_ipfs_publish_blocking, hermes_ipfs_put_dht_value, hermes_ipfs_subscribe_blocking,
-        hermes_ipfs_unpin_file, hermes_ipfs_unsubscribe_blocking,
+        self, blocking, hermes_ipfs_add_file, hermes_ipfs_content_validate,
+        hermes_ipfs_dht_get_providers, hermes_ipfs_dht_provide, hermes_ipfs_evict_peer,
+        hermes_ipfs_get_dht_value, hermes_ipfs_get_file, hermes_ipfs_pin_file,
+        hermes_ipfs_put_dht_value, hermes_ipfs_unpin_file,
     },
     runtime_context::HermesRuntimeContext,
     runtime_extensions::bindings::hermes::ipfs::api::{
@@ -87,7 +86,7 @@ impl Host for HermesRuntimeContext {
     }
 
     fn get_peer_id(&mut self) -> wasmtime::Result<Result<PeerId, Errno>> {
-        let maybe_identity = hermes_ipfs_get_peer_identity_blocking(None)?;
+        let maybe_identity = blocking::hermes_ipfs_get_peer_identity(None)?;
         match maybe_identity {
             Some(identity) => Ok(Ok(identity.peer_id.to_string())),
             None => Ok(Err(Errno::GetPeerIdError)),
@@ -99,7 +98,7 @@ impl Host for HermesRuntimeContext {
         topic: PubsubTopic,
         message: MessageData,
     ) -> wasmtime::Result<Result<(), Errno>> {
-        Ok(hermes_ipfs_publish_blocking(
+        Ok(blocking::hermes_ipfs_publish(
             self.app_name(),
             &topic,
             message,
@@ -110,7 +109,7 @@ impl Host for HermesRuntimeContext {
         &mut self,
         topic: PubsubTopic,
     ) -> wasmtime::Result<Result<bool, Errno>> {
-        Ok(hermes_ipfs_subscribe_blocking(
+        Ok(blocking::hermes_ipfs_subscribe(
             ipfs::SubscriptionKind::Default,
             self.app_name(),
             None,
@@ -123,7 +122,7 @@ impl Host for HermesRuntimeContext {
         &mut self,
         topic: PubsubTopic,
     ) -> wasmtime::Result<Result<bool, Errno>> {
-        Ok(hermes_ipfs_unsubscribe_blocking(
+        Ok(blocking::hermes_ipfs_unsubscribe(
             ipfs::SubscriptionKind::Default,
             self.app_name(),
             &topic,
